requirements.txt
Package             Version
------------------- ---------
asgiref             3.6.0
certifi             2022.12.7
charset-normalizer  3.0.1
Django              4.1.7
djangorestframework 3.14.0
docopt              0.6.2
idna                3.4
joblib              1.2.0
numpy               1.24.2
pip                 23.0.1
pipreqs             0.4.11
pytz                2022.7.1
requests            2.28.2
scikit-learn        1.2.1
scipy               1.10.1
setuptools          67.1.0
sqlparse            0.4.3
threadpoolctl       3.1.0
urllib3             1.26.14
wheel               0.38.4
yarg                0.1.9
